Added theme-showcase template to use on themes page and in the playground
