+++ apps/docsite/src/components/playgroundLink.ts
@@ -3,9 +3,13 @@
 import {compressToEncodedURIComponent} from 'lz-string';
 
 /**
- * Build a playground URL with prepopulated source code.
+ * Build a playground URL with prepopulated source code, and optionally a
+ * seeded theme. When `theme` is given (a short theme slug like "neutral"),
+ * it's added as a `?theme=` query param so the playground opens with that
+ * theme applied + its Theme editor populated; the code rides in the hash.
  */
-export function buildPlaygroundHref(source: string): string {
+export function buildPlaygroundHref(source: string, theme?: string): string {
   const compressed = compressToEncodedURIComponent(source);
-  return `/playground#code=${compressed}`;
+  const query = theme ? `?theme=${encodeURIComponent(theme)}` : '';
+  return `/playground${query}#code=${compressed}`;
 }
